consolidated 'legacy' classes to one file
